requirements.txt
Flask
python-binance